# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nour — Egyptian Fashion & Beauty Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: tenantId
          in: query
          required: false
          schema:
            type: integer
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/variants:
    get:
      operationId: listProductVariants
      tags: [products]
      summary: List all variants for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of variants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductVariant"
    post:
      operationId: createProductVariant
      tags: [products]
      summary: Add a variant to a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductVariantBody"
      responses:
        "201":
          description: Created variant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
  /products/{id}/variants/{variantId}:
    put:
      operationId: updateProductVariant
      tags: [products]
      summary: Update a product variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: variantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductVariantBody"
      responses:
        "200":
          description: Updated variant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
    delete:
      operationId: deleteProductVariant
      tags: [products]
      summary: Delete a product variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: variantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ProductVariant:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        size:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        colorHex:
          type: string
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
        stock:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, productId, stock, createdAt]
    CreateProductVariantBody:
      type: object
      properties:
        size:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        colorHex:
          type: string
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
        stock:
          type: integer
      required: [stock]
    UpdateProductVariantBody:
      type: object
      properties:
        size:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        colorHex:
          type: string
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
        stock:
          type: integer
    Product:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        tenantName:
          type: string
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        name:
          type: string
        description:
          type: string
          default: ""
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        stock:
          type: integer
        featured:
          type: boolean
        status:
          type: string
          enum: [active, out_of_stock, hidden]
        orderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, tenantName, name, description, price, stock, featured, status, orderCount, createdAt]
    CreateProductBody:
      type: object
      properties:
        tenantId:
          type: integer
        categoryId:
          type: integer
          nullable: true
        name:
          type: string
        description:
          type: string
          default: ""
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        stock:
          type: integer
        featured:
          type: boolean
          default: false
        status:
          type: string
          enum: [active, out_of_stock, hidden]
          default: active
      required: [name, price, stock]
    UpdateProductBody:
      type: object
      properties:
        categoryId:
          type: integer
          nullable: true
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        stock:
          type: integer
        featured:
          type: boolean
        status:
          type: string
          enum: [active, out_of_stock, hidden]
...